rubocop corrections

+++ lib/quipper/rubocop/config.rb
@@ -1,10 +1,11 @@
+# frozen_string_literal: true
+
 require "thor"
 
 module Quipper
   module Rubocop
     module Config
       class CLI < Thor
-
         RUBOCOP_CONFIG_FILE_NAME = ".rubocop.yml"
         RUBOCOP_TODO_FILE_NAME = ".rubocop_todo.yml"
         GITHOOK_FILE_PATH = ".git/hooks/pre-push"
@@ -30,12 +31,8 @@ module Quipper
 
         private
 
-        def template_path
-          File.expand_path("../../../templates", __dir__)
-        end
-
         def create_rubocop_config!(config_file_name)
-          puts "#{File.exist?(config_file_name) ? "overwrite" : "create"} #{config_file_name}"
+          puts "#{File.exist?(config_file_name) ? 'overwrite' : 'create'} #{config_file_name}"
           FileUtils.copy_file(File.join(template_path, config_file_name), config_file_name)
         end
 
@@ -47,24 +44,26 @@ module Quipper
         def create_prepush_hook!(file_path)
           puts "adding git prepush hook at #{file_path}"
 
-          if File.exist?(file_path)
-            answer = ask "A prepush githook file already existas at #{file_path}, would you like to overwrite it? (y/n)"
-            if answer != "y" && answer != "yes"
-              puts "skipping githook creation"
-              return
-            end
-          end
+          raise "File already exists at #{file_path}" if File.exist?(file_path)
 
           FileUtils.copy_file(File.join(template_path, ".githooks/pre-push"), file_path)
-          FileUtils.chmod(0755, file_path)
+          FileUtils.chmod(0o755, file_path)
           puts "hook created!"
         end
 
         def remove_file(file_path)
-          if File.exist?(file_path)
+          if File.exist?(file_path) && added_by_this_gem?(file_path)
             File.delete(file_path)
             puts "Removing #{file_path}"
           end
+        end
+
+        def template_path
+          File.expand_path("../../../templates", __dir__)
+        end
+
+        def added_by_this_gem?(file_name)
+          File.read(file_name).include?("quipper-rubocop-config")
         end
       end
     end
